+++ DQMUtils/RD53GainHistograms.h
@@ -35,7 +35,7 @@ class RD53GainHistograms : public DQMHistogramBase
    , stopValue  (stopValue)
   {}
   
-  void book              (TFile* theOutputFile, const DetectorContainer& theDetectorStructure, std::map<std::string, uint32_t> pSettingsMap) override;
+  void book              (TFile* theOutputFile, const DetectorContainer& theDetectorStructure, std::map<std::string, double> pSettingsMap) override;
   void process           ()                                                                                                                  override;
   bool fill              (std::vector<char>& dataBuffer)                                                                                     override { return false; };
   void reset             (void)                                                                                                              override {};
